[mlir][Vector] Make createWriteOrMaskedWrite vector utility
